tensordict fn
