+++ src/aioway/fn/_data.py
