+++ tests/torch/test_tensors.py
@@ -9,7 +9,7 @@ from torch import Tensor
 
 from aioway import fake
 from aioway._previews import Attr
-from aioway.torch import TensorFn
+from aioway.fn import TensorFn
 
 
 @pytest.fixture
